fix(admin): pin MCP routes and instrumentation off Edge (GAP-335)
Turbopack traces process.exit in instrumentation.ts and Node-only
@revealui/mcp/oauth imports from MCP routes that lacked runtime=nodejs.
Move Node boot into instrumentation-node and pin every MCP route.
