+++ apps/admin/src/app/api/mcp/remote-servers/[server]/call-tool-stream/route.ts
@@ -36,6 +36,8 @@ import {
 } from '@/lib/mcp/remote-server-client';
 import { extractRequestContext } from '@/lib/utils/request-context';
 
+export const runtime = 'nodejs';
+
 const IDENTIFIER_RE = /^[A-Za-z0-9_-]{1,64}$/;
 
 type SseEvent =
