+++ apps/admin/src/app/api/mcp/remote-servers/[server]/call-tool/route.ts
@@ -18,6 +18,8 @@ import {
 } from '@/lib/mcp/remote-server-client';
 import { extractRequestContext } from '@/lib/utils/request-context';
 
+export const runtime = 'nodejs';
+
 const IDENTIFIER_RE = /^[A-Za-z0-9_-]{1,64}$/;
 
 export async function POST(
