+++ apps/admin/src/app/api/mcp/remote-servers/[server]/resource/route.ts
@@ -19,6 +19,8 @@ import {
 } from '@/lib/mcp/remote-server-client';
 import { extractRequestContext } from '@/lib/utils/request-context';
 
+export const runtime = 'nodejs';
+
 const IDENTIFIER_RE = /^[A-Za-z0-9_-]{1,64}$/;
 const MAX_URI_LENGTH = 2048;
